+++ pyproject.toml
@@ -40,14 +40,14 @@ requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
 # defines a workflow for cytomining ecosystem software landscape analysis
-[tool.poe.tasks.cytomining_ecosystem]
+[tool.poe.tasks.cytomining_ecosystem_fall_2023]
 # sets a cwd relative for the shell script below
 cwd = "notebooks/cytomining-ecosystem"
 # defines workflow through sequence of python scripts converted from jupyter notebooks
 shell = """
   python seek-landscape-entries.py &&
   python project-github-metrics.py &&
-  python visualize-landscape.py
+  python visualize-landscape-fall-2023.py
 """
 
 [tool.jupytext]
